refactor/809 - Remove subfolders from navigationLibrary
