+++ src/assets/scripts/client/base/ModelSource/modelSourceClassMap.js
@@ -1,5 +1,5 @@
 import _keys from 'lodash/keys';
-import FixModel from '../../navigationLibrary/Fix/FixModel';
+import FixModel from '../../navigationLibrary/FixModel';
 import SpawnPatternModel from '../../trafficGenerator/SpawnPatternModel';
 
 /**
