+++ src/assets/scripts/client/navigationLibrary/NavigationLibrary.js
@@ -4,8 +4,8 @@ import _forEach from 'lodash/forEach';
 import _isNil from 'lodash/isNil';
 import _map from 'lodash/map';
 import _uniq from 'lodash/uniq';
-import FixCollection from './Fix/FixCollection';
-import ProcedureDefinitionModel from './Procedure/ProcedureDefinitionModel';
+import FixCollection from './FixCollection';
+import ProcedureDefinitionModel from './ProcedureDefinitionModel';
 import StaticPositionModel from '../base/StaticPositionModel';
 import { PROCEDURE_TYPE } from '../constants/routeConstants';
 import { degreesToRadians } from '../utilities/unitConverters';
